requirements.txt
python-telegram-bot==20.3
flask==3.1.0
requests==2.32.3
